# Add shortcut for Grisu3 algorithm.