@mazong1123 2023-04-16
Here is the [benchmark data](https://github.com/dotnet/coreclr/pull/14646#issuecomment-351593345) for .NET with the shortcut. Doing similar benchmark tests in rust would be great.