@fee1-dead 2023-04-16
This can be expected since rustc perf is mostly related to the performance of the compiler. You should consider writing your own benchmarks to see the performance changes.