@mazong1123 2023-04-16
Appears current rust benchmark for Grisu3 does not have numbers to cover the fallback cases. I've added 3 benchmarks: `bench_one_exact_inf`, `bench_trailing_zero_exact_inf` and `bench_halfway_point_exact_inf`. Below are the benchmark results in my local machine:

# Before change
benchmarks:
    num::flt2dec::strategy::grisu::bench_big_exact_12             75.00ns/iter  +/- 4.00ns
    num::flt2dec::strategy::grisu::bench_big_exact_3              35.00ns/iter  +/- 1.00ns
    num::flt2dec::strategy::grisu::bench_big_exact_inf            33.15µs/iter  +/- 5.16µs
    num::flt2dec::strategy::grisu::bench_big_shortest             93.00ns/iter  +/- 5.00ns
    **num::flt2dec::strategy::grisu::bench_halfway_point_exact_inf 475.00ns/iter +/- 25.00ns**
    **num::flt2dec::strategy::grisu::bench_one_exact_inf           476.00ns/iter +/- 20.00ns**
    num::flt2dec::strategy::grisu::bench_small_exact_12           58.00ns/iter  +/- 3.00ns
    num::flt2dec::strategy::grisu::bench_small_exact_3            32.00ns/iter  +/- 1.00ns
    num::flt2dec::strategy::grisu::bench_small_exact_inf           1.00µs/iter +/- 27.00ns
    num::flt2dec::strategy::grisu::bench_small_shortest           51.00ns/iter  +/- 2.00ns
    **num::flt2dec::strategy::grisu::bench_trailing_zero_exact_inf 490.00ns/iter +/- 16.00ns**

# After change
benchmarks:
    num::flt2dec::strategy::grisu::bench_big_exact_12             77.00ns/iter   +/- 3.00ns
    num::flt2dec::strategy::grisu::bench_big_exact_3              35.00ns/iter   +/- 1.00ns
    num::flt2dec::strategy::grisu::bench_big_exact_inf            33.11µs/iter   +/- 2.61µs
    num::flt2dec::strategy::grisu::bench_big_shortest             92.00ns/iter   +/- 7.00ns
    **num::flt2dec::strategy::grisu::bench_halfway_point_exact_inf 413.00ns/iter  +/- 31.00ns**
    **num::flt2dec::strategy::grisu::bench_one_exact_inf           416.00ns/iter  +/- 65.00ns**
    num::flt2dec::strategy::grisu::bench_small_exact_12           62.00ns/iter   +/- 3.00ns
    num::flt2dec::strategy::grisu::bench_small_exact_3            32.00ns/iter   +/- 1.00ns
    num::flt2dec::strategy::grisu::bench_small_exact_inf         985.00ns/iter +/- 319.00ns
    num::flt2dec::strategy::grisu::bench_small_shortest           54.00ns/iter   +/- 8.00ns
    **num::flt2dec::strategy::grisu::bench_trailing_zero_exact_inf 434.00ns/iter  +/- 26.00ns**

| Benchmark  | Before | After | Improvement |
| ------------- | -------------  | -------------  | ------------- |
| (num::flt2dec::strategy::grisu::bench_halfway_point_exact_inf)  | 475.00ns/iter | 413.00ns/iter | 13.05% |
| (num::flt2dec::strategy::grisu::bench_one_exact_inf)  | 476.00ns/iter |  416.00ns/iter | 12.60%|
| num::flt2dec::strategy::grisu::bench_trailing_zero_exact_inf   | 490.00ns/iter | 434.00ns/iter | 11.42% |